+++ query/token/client/thrift/client.go
@@ -26,7 +26,7 @@ func (c client) Ping() string {
 	return reply
 }
 
-func (c client) VerifyToken(key string, ttype int64, overhour float64) int64 {
+func (c client) VerifyToken(key string, ttype int64, overhour int64) int64 {
 	reply, err := c.TokenServiceClient.VerifyToken(key, ttype, overhour)
 	if err != nil {
 		c.Logger.Log("err", err)
