+++ query/token/thrift_binding.go
@@ -12,6 +12,6 @@ func (tb thriftBinding) Ping() (string, error) {
 	return tb.TokenService.Ping(), nil
 }
 
-func (tb thriftBinding) VerifyToken(key string, ttype int64, overhour float64) (int64, error) {
+func (tb thriftBinding) VerifyToken(key string, ttype int64, overhour int64) (int64, error) {
 	return tb.TokenService.VerifyToken(key, ttype, overhour), nil
 }
